Add the first version of sound analyzer
The analyzer works by reading sound data using ALSA,
and then running it through FFT.  The freqency range
is split into frequency bands over which a mean is
calculated.  A rolling weighted average of energy
levels on these bands are kept for smoothing.

Before data is returned it can be scaled, weighted
and clipped.

Signed-off-by: Fredrik Lindberg <fli@shapeshifter.se>