Implement automatic sound level normalization
Adjust the sound level scaling factor automatically based
on the average loudness over the spectrum.

If the average level is above 50% of the max level the factor
is increased and if the average level is below 5% the factor
is decreased.  The average level is a "reverse" EMA that
favours old values instead of new values to get a more stable value.

Signed-off-by: Fredrik Lindberg <fli@shapeshifter.se>