sound: add onset and transient calculation
Initial onset and transient calclation.  The onset point
is defined as the point where the difference between the
current spectral flux and the moving average rectified
spectral flux over the window M is greater than the median
flux over the window M*2.

The transient period is defined as the point from onset
until energy level falls below mean of the attack vector.
Where the attack vector is all the points from onset where
the energy level increases.

Transient is returned as a value [1.0, 0.0] where 1.0
indicates the onset point and 0 the end of the period.
The average transient period is kept and used to calculate
the step size for each frame.

Signed-off-by: Fredrik Lindberg <fli@shapeshifter.se>