Add generic output framework
Replace the old light modules with a more generic output framework
with modules that can be enabled and disabled on-the-fly.

A output module needs to define an Output object that implements
an update method, the constructor is called when the module
is enabled and the update method will be called for each data point.

Any output module named output_*.py will be available if it can
be successfully imported.

Hardwire the main process to the debug output for now.

Signed-off-by: Fredrik Lindberg <fli@shapeshifter.se>