output: allow multiple instances of a output module
This changes the behaivour of enabling/disabling modules
by requiring that that an instance of the module to be
created first. This will allow the same module to be
used multiple times with difference parameters.

Instances are created/destroyed with the commands
output create and output destroy.  These instances
can then be disabled/enabled using the instance name.

 > output create <module>
 instance0 created
 > output enable instance0
 instance0 enabled
 > output disable instance0
 instance0 disabled
 > output destroy instance0
 instance0 destroyed

Available output modules can be displayed with
output modules.  Instances are shown with output list
and active instances with output active.

Signed-off-by: Fredrik Lindberg <fli@shapeshifter.se>