output: add a simple visual output module
Add an _very_ simple graphical visualizer that displays the
level, flux and onset transients for each band as well as
the "raw" spectrum data.  The data is displayed as bar graphs,
the onset detection is represented by a green dot over each band.

This modules is suitable for testing and debugging.

Requires pyglet and pyglet2d.

Signed-off-by: Fredrik Lindberg <fli@shapeshifter.se>