sound: smoother adjustment of transient period
Keep a rolling moving average of the transient period and allow it to
both grow and shrink, but in a more smooth way.

If the level is below 75% of the mean attack vector and the remaning transient
period is 0 or less the transient time is recorded into the RMA filter.
If the level is below but the transient period is greater than 0 the
period is divided in half for each time slot.

If the level doesn't hit 75% of the mean attack vector for 5 transient periods
we just abort the current transient and ignore the period length.

This results in a much better onset detection that more naturally follows
the music since since sub-onsets that happens during the main transient
will be ignored.

Signed-off-by: Fredrik Lindberg <fli@shapeshifter.se>