output: add support for output helper modules
Output helpers are singleton objects that when created
register with the output system.  On update the hooks
on_update_begin() and on_update_end() are called in the
helper to allow it to execute actions.

The purpose of this is to perform aggregation of data and
allow multiple output modules to use the same method to
output data in a more efficient way.

Signed-off-by: Fredrik Lindberg <fli@shapeshifter.se>