output: add configuration support for modules
Output modules can define a config dict that describes any
configuration that is possible to pass to the module at creation.

The configuration definition has the following syntax
  config = {
  	"key" : {
  		"type" : <int,float,str>    # What kind of data to expect
  		"default" : "default_value" # Default value
  		"help" : "help text describing the option"
  	}
  }
All sub-fields are optional. If type is specified the supplied value is
type checked and ignored if its of the wrong type.  If default is supplied
a value is always passed in regardless of if a value was supplied or not.

It's optional for the module to supply a config dict, but the constructor
must always accept a config dict.

Options are passed in from the command line at creation as key-value pairs.
  output create <module> key0=value [key1=value]

Signed-off-by: Fredrik Lindberg <fli@shapeshifter.se>