sound: improve level normalization algorithm
Modify the sound level normalization to use a PID-based algorithm
instead of previous, quite stupid way of calculating the scale factor.

However, a plain PID algorithm would be too good at tracking the input
value and would alter the scale factor in a way that would more or
less result in a constant input value.

To avoid this, but still have a rapid increase to a useful value,
the scaling factor is also kept in a 30 second RMA filter.
Based on the loudness difference from the target level, either
the raw value or the filtered value is used.  For small enough
deltas the value is "locked" from changes all together.

This is should keep the scaling factor "stable" during temporary,
silent/low energy periods in the input.

Signed-off-by: Fredrik Lindberg <fli@shapeshifter.se>